requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
python-multipart
pytesseract
PyPDF2